chore: change how state type is passed to createEffects
The createEffects function needs to get the slice's state.
I tried hard to get it from its functional params:
- the `inputs` param has it in theory, but I could not make it work to
  read it from there
- the `forSlice()` could be changed to accept the entire slice instead
  of only its name, but that creates a circular type reference

So I was passing in the initial state without actually using it in the
function, only to get the State type from it.
However, when using this package in a consumer, the initialState
variables end up being undefined at the time of usage, although only
when consuming the ESM output, not the CJS.

So here I changed to code to get the type argument explicitly, with the
same double-function trick that createEffectInputs is also using to get
one type param explicitly and infer the rest.